Refactor user input form and loading srp
