+++ client/components/Loading/constants.js
@@ -1,0 +1,2 @@
+export const WELCOME_DESCRIPTION =
+  "Welcome to Purleiu.io, an orthogonal projection of boundaries for ride sharing applications. Enter a dollar amount between $10 - $400 into the form above and we'll calculate a matrix displaying how far you can go in your next Lyft ride request!"
