--- client/components/_reusables/text.js
@@ -1,2 +1,0 @@
-export const WELCOME_DESCRIPTION =
-  "Welcome to MoneyMile, a Lyft API ride estimates application. Enter a dollar amount between $10 - $400 into the form above and we'll calculate a matrix displaying how far you can go in your next Lyft ride request!"
