+++ client/components/user/user_ride_selection.jsx
@@ -3,25 +3,15 @@ import React from "react"
 class UserRideSelection extends React.Component {
   constructor(props) {
     super(props)
-    // this.state = {
-    //   selected: ["lyft"]
-    // };
   }
 
   updateType = type => {
-    // console.log("selectedRideTypes", this.props.selectedRideTypes);
     return e => {
-      // if (this.state.selected.includes(type)) {
       if (this.props.selectedRideTypes.includes(type)) {
         e.currentTarget.classList.remove("selected")
-        // const index = selectedRideTypes.indexOf(type);
-        // selectedRideTypes.splice(index, 1);
-        // this.setState({ selected: selectedRideTypes });
         this.props.clearOverlay(type)
       } else {
         e.currentTarget.classList.add("selected")
-        // selectedRideTypes.push(type);
-        // this.setState({selected: selectedRideTypes});
         this.props.getRideType(type)
       }
     }
